requirements.txt
discord.py
ollama
python-dotenv
sentence_transformers
langchain
chardet
numpy